+++ src/onyx/log/zookeeper.clj
@@ -1,7 +1,7 @@
 (ns onyx.log.zookeeper
   (:require [clojure.core.async :refer [chan >!! <!! close! thread]]
             [com.stuartsierra.component :as component]
-            [taoensso.timbre :refer [fatal]]
+            [taoensso.timbre :refer [fatal warn]]
             [zookeeper :as zk]
             [onyx.extensions :as extensions]
             [onyx.compression.nippy :refer [compress decompress]])
@@ -24,6 +24,9 @@
 (defn workflow-path [prefix]
   (str (prefix-path prefix) "/workflow"))
 
+(defn flow-path [prefix]
+  (str (prefix-path prefix) "/flow"))
+
 (defn task-path [prefix]
   (str (prefix-path prefix) "/task"))
 
@@ -33,9 +36,12 @@
 (defn origin-path [prefix]
   (str (prefix-path prefix) "/origin"))
 
+(defn job-scheduler-path [prefix]
+  (str (prefix-path prefix) "/job-scheduler"))
+
 (defn initialize-origin! [conn config prefix]
   (let [node (str (origin-path prefix) "/origin")
-        bytes (compress {:message-id -1 :replica {:job-scheduler (:onyx.peer/job-scheduler config)}})]
+        bytes (compress {:message-id -1 :replica {}})]
     (zk/create conn node :data bytes :persistent? true)))
 
 (defrecord ZooKeeper [config]
@@ -52,9 +58,11 @@
       (zk/create conn (log-path onyx-id) :persistent? true)
       (zk/create conn (catalog-path onyx-id) :persistent? true)
       (zk/create conn (workflow-path onyx-id) :persistent? true)
+      (zk/create conn (flow-path onyx-id) :persistent? true)
       (zk/create conn (task-path onyx-id) :persistent? true)
       (zk/create conn (sentinel-path onyx-id) :persistent? true)
       (zk/create conn (origin-path onyx-id) :persistent? true)
+      (zk/create conn (job-scheduler-path onyx-id) :persistent? true)
 
       (initialize-origin! conn config onyx-id)
       (assoc component :server server :conn conn :prefix onyx-id)))
@@ -90,8 +98,9 @@
 (defmethod extensions/read-log-entry ZooKeeper
   [{:keys [conn opts prefix] :as log} position]
   (let [node (str (log-path prefix) "/entry-" (pad-sequential-id position))
-        content (decompress (:data (zk/data conn node)))]
-    (assoc content :message-id position)))
+        data (zk/data conn node)
+        content (decompress (:data data))]
+    (assoc content :message-id position :created-at (:ctime (:stat data)))))
 
 (defmethod extensions/register-pulse ZooKeeper
   [{:keys [conn opts prefix] :as log} id]
@@ -110,14 +119,28 @@
         ;; Node doesn't exist.
         (>!! ch true)))))
 
+(defn find-job-scheduler [log]
+  (loop []
+    (if-let [chunk
+             (try
+               (extensions/read-chunk log :job-scheduler nil)
+               (catch Exception e
+                 (warn e)
+                 (warn "Job scheduler couldn't be discovered. Backing off 500ms and trying again...")
+                 nil))]
+      (:job-scheduler chunk)
+      (do (Thread/sleep 500)
+          (recur)))))
+
 (defmethod extensions/subscribe-to-log ZooKeeper
   [{:keys [conn opts prefix] :as log} ch]
   (let [rets (chan)]
     (thread
      (try
-       (let [origin (extensions/read-chunk log :origin nil)
+       (let [job-scheduler (find-job-scheduler log)
+             origin (extensions/read-chunk log :origin nil)
              starting-position (inc (:message-id origin))]
-         (>!! rets (:replica origin))
+         (>!! rets (assoc (:replica origin) :job-scheduler job-scheduler))
          (close! rets)
          (loop [position starting-position]
            (let [path (str (log-path prefix) "/entry-" (pad-sequential-id position))]
@@ -139,6 +162,7 @@
                      (recur)))))
              (recur (inc position)))))
        (catch org.apache.zookeeper.KeeperException$ConnectionLossException e
+         ;; ZooKeeper has been shutdown, close the subscriber cleanly.
          (close! ch))
        (catch org.apache.zookeeper.KeeperException$SessionExpiredException e
          (close! ch))
@@ -158,6 +182,12 @@
         bytes (compress chunk)]
     (zk/create conn node :persistent? true :data bytes)))
 
+(defmethod extensions/write-chunk [ZooKeeper :flow-conditions]
+  [{:keys [conn opts prefix] :as log} kw chunk id]
+  (let [node (str (flow-path prefix) "/" id)
+        bytes (compress chunk)]
+    (zk/create conn node :persistent? true :data bytes)))
+
 (defmethod extensions/write-chunk [ZooKeeper :task]
   [{:keys [conn opts prefix] :as log} kw chunk id]
   (let [node (str (task-path prefix) "/" (:id chunk))
@@ -170,6 +200,12 @@
         bytes (compress chunk)]
     (zk/create conn node :persistent? true :data bytes)))
 
+(defmethod extensions/write-chunk [ZooKeeper :job-scheduler]
+  [{:keys [conn opts prefix] :as log} kw chunk id]
+  (let [node (str (job-scheduler-path prefix) "/scheduler")
+        bytes (compress chunk)]
+    (zk/create conn node :persistent? true :data bytes)))
+
 (defmethod extensions/read-chunk [ZooKeeper :catalog]
   [{:keys [conn opts prefix] :as log} kw id]
   (let [node (str (catalog-path prefix) "/" id)]
@@ -178,6 +214,11 @@
 (defmethod extensions/read-chunk [ZooKeeper :workflow]
   [{:keys [conn opts prefix] :as log} kw id]
   (let [node (str (workflow-path prefix) "/" id)]
+    (decompress (:data (zk/data conn node)))))
+
+(defmethod extensions/read-chunk [ZooKeeper :flow-conditions]
+  [{:keys [conn opts prefix] :as log} kw id]
+  (let [node (str (flow-path prefix) "/" id)]
     (decompress (:data (zk/data conn node)))))
 
 (defmethod extensions/read-chunk [ZooKeeper :task]
@@ -193,6 +234,11 @@
 (defmethod extensions/read-chunk [ZooKeeper :origin]
   [{:keys [conn opts prefix] :as log} kw id]
   (let [node (str (origin-path prefix) "/origin")]
+    (decompress (:data (zk/data conn node)))))
+
+(defmethod extensions/read-chunk [ZooKeeper :job-scheduler]
+  [{:keys [conn opts prefix] :as log} kw id]
+  (let [node (str (job-scheduler-path prefix) "/scheduler")]
     (decompress (:data (zk/data conn node)))))
 
 (defmethod extensions/update-origin! ZooKeeper