+++ src/onyx/peer/task_lifecycle.clj
@@ -5,7 +5,7 @@
               [taoensso.timbre :refer [info warn trace] :as timbre]
               [onyx.log.commands.common :as common]
               [onyx.log.entry :as entry]
-              [onyx.static.planning :refer [find-task]]
+              [onyx.static.planning :refer [find-task build-pred-fn]]
               [onyx.messaging.acking-daemon :as acker]
               [onyx.peer.task-lifecycle-extensions :as l-ext]
               [onyx.peer.pipeline-extensions :as p-ext]
@@ -96,17 +96,36 @@
     (repeat (count segments) nil)
     (map (fn [x] (acker/gen-ack-value)) segments)))
 
-(defn with-clean-up [f ch dead-ch release-f exception-f]
-  (try
-    (f)
-    (catch Exception e
-      (exception-f e)
-      (close! ch)
-      ;; Unblock any blocked puts
-      (<!! ch))
-    (finally
-     (>!! dead-ch true)
-     (release-f))))
+(defn join-output-paths [all to-add downstream]
+  (cond (= to-add :all) (into #{} downstream)
+        (= to-add :none) #{}
+        :else (clojure.set/union (into #{} all) (into #{} to-add))))
+
+(defn choose-output-paths [flow-conditions event new downstream]
+  (if (seq flow-conditions)
+    (reduce
+     (fn [{:keys [paths exclusions] :as all} entry]
+       (if ((:flow/predicate entry) [event new])
+         (if (:flow/short-circuit? entry)
+           (reduced {:paths (join-output-paths paths (:flow/to entry) downstream)
+                     :exclusions (clojure.set/union (into #{} exclusions) (into #{} (:flow/exclude-keys entry)))})
+           {:paths (join-output-paths paths (:flow/to entry) downstream)
+            :exclusions (clojure.set/union (into #{} exclusions) (into #{} (:flow/exclude-keys entry)))})
+         all))
+     {:paths #{} :exclusions #{}}
+     flow-conditions)
+    {:paths downstream}))
+
+(defn output-paths
+  [{:keys [onyx.core/serialized-task onyx.core/compiled-flow-conditions] :as event}]
+  (merge
+   event
+   {:onyx.core/result-paths
+    (map
+     (fn [segment]
+       (choose-output-paths compiled-flow-conditions event segment
+                            (keys (:egress-queues serialized-task))))
+     (:onyx.core/results event))}))
 
 (defn inject-batch-resources [event]
   (let [cycle-params {:onyx.core/lifecycle-id (java.util.UUID/randomUUID)}]
@@ -191,6 +210,16 @@
     (let [entry (entry/create-log-entry :kill-job {:job job-id})]
       (>!! outbox-ch entry))))
 
+(defn only-relevant-branches [flow task]
+  (filter #(= (:flow/from %) task) flow))
+
+(defn compile-flow-conditions [flow-conditions task-name]
+  (let [conditions (only-relevant-branches flow-conditions task-name)]
+    (map
+     (fn [condition]
+       (assoc condition :flow/predicate (build-pred-fn (:flow/predicate condition) condition)))
+          conditions)))
+
 (defn run-task-lifecycle [init-event kill-ch]
   (loop [event init-event]
     (when (first (alts!! [kill-ch] :default true))
@@ -222,6 +251,7 @@
     (try
       (let [catalog (extensions/read-chunk log :catalog job-id)
             task (extensions/read-chunk log :task task-id)
+            flow-conditions (extensions/read-chunk log :flow-conditions job-id)
             catalog-entry (find-task catalog (:name task))
 
             kill-ch (chan (dropping-buffer 1))
@@ -234,6 +264,8 @@
                            :onyx.core/task (:name task)
                            :onyx.core/catalog catalog
                            :onyx.core/workflow (extensions/read-chunk log :workflow job-id)
+                           :onyx.core/flow-conditions flow-conditions
+                           :onyx.core/compiled-flow-conditions (compile-flow-conditions flow-conditions (:name task))
                            :onyx.core/task-map catalog-entry
                            :onyx.core/serialized-task task
                            :onyx.core/params (resolve-calling-params catalog-entry  opts)