+++ src/onyx/static/planning.clj
@@ -1,5 +1,6 @@
 (ns onyx.static.planning
-  (:require [com.stuartsierra.dependency :as dep])
+  (:require [com.stuartsierra.dependency :as dep]
+            [onyx.peer.operation :refer [kw->fn]])
   (:import [java.util UUID]))
 
 (defmulti create-io-task
@@ -88,4 +89,33 @@
           (concat tasks (create-task catalog element parent-entries children))))
       []
       sorted-dag))))
+
+(defn pred-fn? [expr]
+  (and (keyword? expr)
+       (not= :and expr)
+       (not= :or expr)
+       (not= :not expr)))
+
+(defn build-pred-fn [expr entry]
+  (if (pred-fn? expr)
+    (fn [xs] (apply (kw->fn expr) xs))
+    (let [[op & more :as full-expr] expr]
+      (cond (= op :and)
+            (do (assert (> (count more) 1) ":and takes at least two predicates")
+                (fn [xs]
+                  (every? identity (map (fn [token] ((build-pred-fn token entry) xs)) more))))
+
+            (= op :or)
+            (do (assert (> (count more) 1) ":or takes at least two predicates")
+                (fn [xs]
+                  (some identity (map (fn [token] ((build-pred-fn token entry) xs)) more))))
+
+            (= op :not)
+            (do (assert (= 1 (count more)) ":not only takes one predicate")
+                (fn [xs]
+                  (not ((build-pred-fn (first more) entry) xs))))
+
+            :else
+            (fn [xs]
+              (apply (kw->fn op) (concat xs (map (fn [arg] (get entry arg)) more))))))))